Alternative mobile menu
I discovered some strange behavior with alpinejs's x-show and x-data. If
the component to be shown has an id set then x-show uses the last set
value of x-data rather than the value sent in the html response.

I'm not sure why this is. I am going to try a few more experiments with
the alpine-morph htmx extension, to see if I prefer that approach. I
think that will allow me to easily keep the menu open after clicking a
navigation link.